Rename code files without underscores.
